Fix build process running out of memory
I think what was going on is that the whole directory was included on
every build, including the `target` directory. After a few builds the
link/build time blew up.

So this only includes the `templates` subdirectory which makes build a
lot faster and should help with the memory issues.

Co-authored-by: Bennet <bennet@zed.dev>